fix: Route selectConnection results through ApiResult + showMessages
